+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- Fetching non-existent or dynamic properties no longer calls refresh().
+
 ## 0.6 - 2017-12-02
 ### Changed
 - Support Symfony 4.